[DL-17153] Address jackson-core vulnerability
