+++ it/test/endpoints/GetDataSpec.scala
@@ -14,43 +14,43 @@
  * limitations under the License.
  */
 
-package tests.endpoints
+package endpoints
 
 import play.api.libs.json.{JsValue, Json}
 import play.api.libs.ws.WSResponse
-import tests.helpers.IntegrationBaseSpec
-import tests.stubs.AuthStub
+import helpers.IntegrationBaseSpec
+import stubs.AuthStub
 
-class StoreDataSpec extends IntegrationBaseSpec {
+class GetDataSpec extends IntegrationBaseSpec {
 
   val path = "/data/123456789/testKey"
-  val invalidPath = "/data/123456789/testKey2"
+  val invalidPath = "/data/12345678901/testKey2"
 
   val body: JsValue = Json.parse("""{"forename" : "john"}""")
 
-  "PUT /data/:vrn/:key" when {
+  "GET /data/:vrn/:key" when {
 
     "the user is authorised" when {
 
-      "mongo successful receives the sent data" should {
+      "mongo successful receives the data" should {
 
-        "get a 204 response and no body is returned" in {
+        "get a 200 response and the correct JSON body" in {
 
           AuthStub.authorised()
-          val response: WSResponse = put(path)(body)
-          response.status shouldBe 204
-          response.body shouldBe ""
+          put(path)(body)
+          val response: WSResponse = get(path)
+          response.status shouldBe 200
+          response.body shouldBe "{\"forename\":\"john\"}"
 
         }
-
       }
-      "user is trying to add a currently existing record" should {
+      "user is unsuccessful" should {
 
-        "get a 204 response and no body returned" in {
+        "get a 404 response and the correct error message" in {
 
-          val response: WSResponse = put(invalidPath)(body)
-          response.status shouldBe 204
-          response.body shouldBe ""
+          val response: WSResponse = get(invalidPath)
+          response.status shouldBe 404
+          response.body shouldBe "{\"message\":\"No data found for vrn: 12345678901 and key: testKey2\"}"
 
         }
       }
@@ -58,10 +58,10 @@ class StoreDataSpec extends IntegrationBaseSpec {
 
     "user is unauthorised" should {
 
-      "get a 401 response and no body is returned " in {
+      "get a 401 response and no body is returned" in {
 
         AuthStub.unauthenticated()
-        val response: WSResponse = put(path)(body)
+        val response: WSResponse = get(path)
         response.status shouldBe 401
         response.body shouldBe ""
 
@@ -74,11 +74,12 @@ class StoreDataSpec extends IntegrationBaseSpec {
       "get a 403 response and no body is returned" in {
 
         AuthStub.forbidden()
-        val response: WSResponse = put(path)(body)
+        val response: WSResponse = get(path)
         response.status shouldBe 403
         response.body shouldBe ""
 
       }
+
     }
   }
 }
