+++ project/plugins.sbt
@@ -20,8 +20,8 @@ resolvers += "Typesafe Releases" at "https://repo.typesafe.com/typesafe/releases
 
 addSbtPlugin("uk.gov.hmrc"       % "sbt-auto-build"        % "3.24.0")
 addSbtPlugin("uk.gov.hmrc"       % "sbt-distributables"    % "2.6.0")
-addSbtPlugin("org.playframework" % "sbt-plugin"            % "3.0.6")
-addSbtPlugin("org.scoverage"     % "sbt-scoverage"         % "2.3.1")
+addSbtPlugin("org.playframework" % "sbt-plugin"            % "3.0.9")
+addSbtPlugin("org.scoverage"     % "sbt-scoverage"         % "2.4.0")
 addSbtPlugin("org.scalastyle"   %% "scalastyle-sbt-plugin" % "1.0.0")
 
 ThisBuild / libraryDependencySchemes += "org.scala-lang.modules" %% "scala-xml" % VersionScheme.Always
