+++ project/scaffold.sbt
@@ -14,4 +14,4 @@
  * limitations under the License.
  */
 
-addSbtPlugin("org.foundweekends.giter8" % "sbt-giter8-scaffold" % "0.12.0")
+addSbtPlugin("org.foundweekends.giter8" % "sbt-giter8-scaffold" % "0.17.0")
